# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Infona",
    "description": "Living Knowledge Graph Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/graphs/{tenant}/functions": {
      "post": {
        "tags": [
          "functions"
        ],
        "summary": "Register Function",
        "operationId": "register_function_graphs__tenant__functions_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FunctionRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "functions"
        ],
        "summary": "List Functions",
        "operationId": "list_functions_graphs__tenant__functions_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FunctionRef"
                  },
                  "title": "Response List Functions Graphs  Tenant  Functions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graphs/{tenant}/ask": {
      "post": {
        "tags": [
          "ask"
        ],
        "summary": "Ask Question",
        "operationId": "ask_question_graphs__tenant__ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NLQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NLResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FunctionRef": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "endpoint_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint Url"
          },
          "tier": {
            "$ref": "#/components/schemas/FunctionTier",
            "default": "custom"
          }
        },
        "type": "object",
        "required": [
          "name",
          "entity_type"
        ],
        "title": "FunctionRef"
      },
      "FunctionRegister": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "entity_type": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Entity Type"
          },
          "endpoint_url": {
            "type": "string",
            "title": "Endpoint Url",
            "description": "HTTPS endpoint for the function"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "entity_type",
          "endpoint_url"
        ],
        "title": "FunctionRegister"
      },
      "FunctionTier": {
        "type": "string",
        "enum": [
          "platform",
          "custom"
        ],
        "title": "FunctionTier"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NLQuery": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "kg_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kg Name",
            "description": "Query a specific knowledge graph"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Override the query generation model (OpenRouter model ID)"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "NLQuery"
      },
      "NLResult": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sparql": {
            "type": "string",
            "title": "Sparql"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "ontology": {
            "type": "string",
            "title": "Ontology",
            "description": "Ontology summary text passed to the LLM for SPARQL generation",
            "default": ""
          },
          "functions_invoked": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Functions Invoked"
          },
          "timing": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "title": "Timing",
            "description": "Stage latencies in ms and metadata"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sparql",
          "explanation"
        ],
        "title": "NLResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}